+++ src/pl-data.h
@@ -311,7 +311,7 @@ and while loading .wic files.  It comes at no price.
 #define functorHashValue(f, n)	((f)>>(LMASK_BITS) & ((n)-1))
 #define indexFunctor(w)	((w)>>(LMASK_BITS+F_ARITY_BITS))
 #define valueFunctor(w) fetchFunctorArray(indexFunctor(w))
-#define _arityFunc_(w)	((int)(((w) >> LMASK_BITS) & F_ARITY_MASK))
+#define _arityFunc_(w)	((size_t)(((w) >> LMASK_BITS) & F_ARITY_MASK))
 #define arityFunctor(w) (unlikely(_arityFunc_(w) == F_ARITY_MASK) \
 				? valueFunctor(w)->arity \
 				: _arityFunc_(w) )
